Fixed up voting
Moved secured /idea/ resources to a different module. This allowed reduced code repetition as we can assume it can only hit those actions if a user is logged in
