+++ src/Ideastrike.Nancy/Ideastrike.Nancy.csproj
@@ -162,6 +162,7 @@
     <Compile Include="Modules\ApiModule.cs" />
     <Compile Include="Modules\CommentModule.cs" />
     <Compile Include="Modules\FeatureModule.cs" />
+    <Compile Include="Modules\IdeaModule.Secured.cs" />
     <Compile Include="Modules\UserModule.cs" />
     <Compile Include="Modules\LoginModule.cs" />
     <Compile Include="Modules\HomeModule.cs" />
